+++ exporter/opentelemetry-exporter-otlp-proto-common/src/opentelemetry/exporter/otlp/proto/common/_internal/metrics_encoder/__init__.py
@@ -252,6 +252,7 @@ def _encode_metric(metric, pb2_metric):
                 attributes=_encode_attributes(data_point.attributes),
                 time_unix_nano=data_point.time_unix_nano,
                 exemplars=_encode_exemplars(data_point.exemplars),
+                flags=int(data_point.flags),
             )
             if isinstance(data_point.value, int):
                 pt.as_int = data_point.value
@@ -272,6 +273,7 @@ def _encode_metric(metric, pb2_metric):
                 explicit_bounds=data_point.explicit_bounds,
                 max=data_point.max,
                 min=data_point.min,
+                flags=int(data_point.flags),
             )
             pb2_metric.histogram.aggregation_temporality = (
                 metric.data.aggregation_temporality
@@ -285,6 +287,7 @@ def _encode_metric(metric, pb2_metric):
                 start_time_unix_nano=data_point.start_time_unix_nano,
                 time_unix_nano=data_point.time_unix_nano,
                 exemplars=_encode_exemplars(data_point.exemplars),
+                flags=int(data_point.flags),
             )
             if isinstance(data_point.value, int):
                 pt.as_int = data_point.value
@@ -328,9 +331,9 @@ def _encode_metric(metric, pb2_metric):
                 zero_count=data_point.zero_count,
                 positive=positive,
                 negative=negative,
-                flags=data_point.flags,
                 max=data_point.max,
                 min=data_point.min,
+                flags=int(data_point.flags),
             )
             pb2_metric.exponential_histogram.aggregation_temporality = (
                 metric.data.aggregation_temporality